move A32NX_Util.js to .ts and /shared
